sports_league/requirements.txt
-i https://pypi.org/simple
asgiref==3.8.1; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
cffi==1.16.0; platform_python_implementation != 'PyPy'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
cryptography==42.0.7
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
dj-database-url==2.2.0
dj-rest-auth[with_social]==6.0.0; python_version >= '3.8'
django==3.2; python_version >= '3.6'
django-allauth==0.61.1
django-cors-headers==4.3.1; python_version >= '3.8'
django-extensions==3.2.3; python_version >= '3.6'
django-filter==23.5; python_version >= '3.7'
djangorestframework==3.15.1; python_version >= '3.6'
djangorestframework-simplejwt==5.3.1; python_version >= '3.8'
drf-yasg==1.21.7; python_version >= '3.6'
graphviz==0.20.3; python_version >= '3.8'
idna==3.7; python_version >= '3.5'
importlib-metadata==7.1.0; python_version < '3.10'
inflection==0.5.1; python_version >= '3.5'
markdown==3.6; python_version >= '3.8'
oauthlib==3.2.2; python_version >= '3.6'
packaging==24.0; python_version >= '3.7'
pillow==10.3.0; python_version >= '3.8'
psycopg2-binary==2.9.9; python_version >= '3.7'
pycparser==2.22; python_version >= '3.8'
pyjwt[crypto]==2.8.0; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
python3-openid==3.2.0
pytz==2024.1
pyyaml==6.0.1; python_version >= '3.6'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
sqlparse==0.5.0; python_version >= '3.8'
typing-extensions==4.12.0; python_version >= '3.8'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==2.2.1; python_version >= '3.8'
zipp==3.19.0; python_version >= '3.8'
